+++ src/metadata/libexif_handler.cc
@@ -183,7 +183,7 @@ void LibExifHandler::process_ifd(ExifContent* content, const std::shared_ptr<Cds
     std::array<char, BUFLEN> exif_entry_buffer;
 #define exif_egv(arg) exif_entry_get_value(arg, exif_entry_buffer.data(), BUFLEN)
 
-    for (unsigned int i = 0; i < content->count; i++) {
+    for (size_t i = 0; i < content->count; i++) {
         ExifEntry* e = content->entries[i];
 
         // log_debug("Processing entry: {}", i);
